fix: incorrect uint64_t cast in bitmask limit validation
The LIMIT_FORMAT_BITMASK validation uses deUint32 for checking the
limit value but incorrectly casts to deUint64 when logging the actual
value. This causes incorrect values to be displayed in the logs.

This bug was originally introduced in c82c0fa6ca (2016) and persisted
through the de* type removal in 1a09796462 (2023).

+++ external/vulkancts/modules/vulkan/api/vktApiFeatureInfo.cpp
@@ -542,7 +542,7 @@ bool validateFeatureLimits(VkPhysicalDeviceProperties *properties, VkPhysicalDev
                 {
                     log << TestLog::Message << "limit validation failed, " << featureLimitTable[ndx].name
                         << " not valid-limit type bitmask actual is "
-                        << *((uint64_t *)((uint8_t *)limits + featureLimitTable[ndx].offset)) << TestLog::EndMessage;
+                        << *((uint32_t *)((uint8_t *)limits + featureLimitTable[ndx].offset)) << TestLog::EndMessage;
                     limitsOk = false;
                 }
             }
